backend/requirements.txt
Flask==3.1.2
Flask-CORS==4.0.0
openai==1.102.0
python-dotenv==1.1.0